+++ CHANGES.rst
@@ -3,6 +3,7 @@
 
 * Set Cache-Control header to 'private' to prevent proxy caching of
   logged-in pages.
+* requires_permission and has_permission now accept multiple permissions.
 
 0.3.13
 ------